\name{categories}
\alias{docs_categories}
\title{Document categories}
\description{Get Document Categories and Subcategories}
\usage{
docs_categories(category = NULL, subs = FALSE, ...)
}
\arguments{
  \item{category}{Optionally, a numeric value containing a Scribd document category.}
  \item{subs}{A logical indicating whether to return subcategory information.}
  \item{...}{Other arguments passed to HTTP request functions.}
}
%\details{}
%\value{}
\references{
\href{http://www.scribd.com/developers/platform/api/docs_getcategories}{API Reference: getCategories}
}
\author{Thomas J. Leeper}
%\seealso{}
%\examples{}
