\name{rscribd-package}
\alias{rscribd-package}
\docType{package}
\title{rscribd API Client Package for R}
\description{This package allows users to publish and retrieve documents from the \href{http://www.scribd.com/}{Scribd} publishing platform.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://www.scribd.com/}{Scribd}

\href{http://www.scribd.com/developers}{API Documentation}

\href{http://www.scribd.com/developers/platform/api}{API Endpoints}
}
\keyword{ package }
%\seealso{}
%\examples{}
